refactor: move the components folder from app to src
