+++ carrerait-roadmap/src/components/Survey/AspirantSurveyResults.tsx
